Radar and ARLayer on one SurfaceView

+++ airspy/src/main/java/com/mac/airspy/utils/ResolutionUtils.java
@@ -6,7 +6,7 @@ import android.content.Context;
  * Created by Maciej on 2014-10-17.
  */
 public class ResolutionUtils {
-    public static int dpToPx(Context ctx, int dp) {
-        return (int) (ctx.getResources().getDisplayMetrics().density * dp);
+    public static float dpToPx(Context ctx, int dp) {
+        return ctx.getResources().getDisplayMetrics().density * dp;
     }
 }
